iitmproject/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
requests
pillow
beautifulsoup4
sqlite-utils
tiktoken
